Implement guest-driven context tracking for the perfetto log backend.
The guest OS can use reserved event NOPs to inform the perfetto backend
about context changes (e.g. a context switch or a compartment switch).
The perfetto backend allocates a separate track to each context and
emits events to the active track in each CPU.
Note that the notion of context switch is entirely given by the guest OS
and it is responsible for correctly updating the state on the backend.
The probe effect in the code due to NOPs instructions should be negligible.